+++ packages/browser/src/tracing/browserTracingIntegration.ts
@@ -236,6 +236,7 @@ export interface BrowserTracingOptions {
   _experiments: Partial<{
     enableInteractions: boolean;
     enableStandaloneClsSpans: boolean;
+    enableStandaloneLcpSpans: boolean;
   }>;
 
   /**
@@ -301,7 +302,7 @@ export const browserTracingIntegration = ((_options: Partial<BrowserTracingOptio
     enableInp,
     enableLongTask,
     enableLongAnimationFrame,
-    _experiments: { enableInteractions, enableStandaloneClsSpans },
+    _experiments: { enableInteractions, enableStandaloneClsSpans, enableStandaloneLcpSpans },
     beforeStartSpan,
     idleTimeout,
     finalTimeout,
@@ -358,6 +359,7 @@ export const browserTracingIntegration = ((_options: Partial<BrowserTracingOptio
         _collectWebVitals?.();
         addPerformanceEntries(span, {
           recordClsOnPageloadSpan: !enableStandaloneClsSpans,
+          recordLcpOnPageloadSpan: !enableStandaloneLcpSpans,
           ignoreResourceSpans,
           ignorePerformanceApiSpans,
         });
@@ -400,7 +402,10 @@ export const browserTracingIntegration = ((_options: Partial<BrowserTracingOptio
     setup(client) {
       registerSpanErrorInstrumentation();
 
-      _collectWebVitals = startTrackingWebVitals({ recordClsStandaloneSpans: enableStandaloneClsSpans || false });
+      _collectWebVitals = startTrackingWebVitals({
+        recordClsStandaloneSpans: enableStandaloneClsSpans || false,
+        recordLcpStandaloneSpans: enableStandaloneLcpSpans || false,
+      });
 
       if (enableInp) {
         startTrackingINP();