+++ packages/cloudflare/src/utils/commonjs.ts
@@ -1,0 +1,8 @@
+/** Detect CommonJS. */
+export function isCjs(): boolean {
+  try {
+    return typeof module !== 'undefined' && typeof module.exports !== 'undefined';
+  } catch {
+    return false;
+  }
+}